V2

+++ page.php
@@ -7,15 +7,18 @@
  * and that other 'pages' on your WordPress site may use a
  * different template.
  *
- * @package starter
+ * @link https://developer.wordpress.org/themes/basics/template-hierarchy/
+ *
+ * @package WordPress
+ * @subpackage starter
+ * @since starter 1.0
  */
 
-get_header();
-?>
+get_header(); ?>
 
+<div class="content_wrapper pt-5 pb-5" role="main">
+	<div class="container">
 
-<div class="content_wrapper" role="main">
-	<div class="container service_page">
 		<!-- breadcrumb -->
 		<?php
 		if ( function_exists( 'yoast_breadcrumb' ) ) {
@@ -23,31 +26,39 @@ get_header();
 		}
 		?>
 		<!-- END breadcrumb -->
+
 		<?php
 		while ( have_posts() ) {
 			the_post();
 			the_title( '<h1 class="entry-title">', '</h1>' );
 
-			switch ( $post->ID ) {
-				case wc_get_page_id( 'cart' ):
-					echo "<div class='cart_page'>" . do_shortcode( '[woocommerce_cart]' ) . '</div>';
-					break;
-				case wc_get_page_id( 'checkout' ):
-					echo "<div class='checkout_page'>" . do_shortcode( '[woocommerce_checkout]' ) . '</div>';
-					break;
-				case wc_get_page_id( 'myaccount' ):
-					echo "<div class='account_page'>" . do_shortcode( '[woocommerce_my_account]' ) . '</div>';
-					break;
-				default:
-					echo '<div class="mt-5">';
-					the_content();
-					echo '</div>';
+			if ( class_exists( 'WooCommerce' ) ) {
+				/*if woocommerce enabled*/
+				switch ( $post->ID ) {
+					case wc_get_page_id( 'cart' ):
+						echo "<div class='cart_page'>" . do_shortcode( '[woocommerce_cart]' ) . '</div>';
+						break;
+					case wc_get_page_id( 'checkout' ):
+						echo "<div class='checkout_page'>" . do_shortcode( '[woocommerce_checkout]' ) . '</div>';
+						break;
+					case wc_get_page_id( 'myaccount' ):
+						echo "<div class='account_page'>" . do_shortcode( '[woocommerce_my_account]' ) . '</div>';
+						break;
+					default:
+						echo '<div class="mt-5">';
+						the_content();
+						echo '</div>';
+				}
+				/*END if woocommerce enabled*/
+			} else {
+				echo '<div class="mt-5">';
+				the_content();
+				echo '</div>';
 			}
 		}
 		?>
 	</div>
 </div><!-- .content_wrapper -->
-
 
 <?php
 get_footer();
